requirements.txt
blinker==1.9.0
branca==0.8.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coordinates==0.4.0
Flask==3.1.2
folium==0.20.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-trace==1.16.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
m3u8==6.0.0
MarkupSafe==3.0.3
mtrpacket==1.0.1
numpy==2.3.4
playwright==1.55.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xyzservices==2025.4.0
